refactor(isotopologues): Create submodule
Moved get_dataframes into its own submodule, with the intent of refactoring it further.
No tests for this function yet.
